entire face RoI
Now we can use the entire face (excluding, mouth, eyes, and eyebrows).
Facial hair is not removed: it is too hard to do for every race and every lighting condition. Maybe in the future.

Now the heavy lifting is done in functions.py. (That way it can be tested part by part).