Worker
Now the data flow graph is created with OOP: a Worker object is created for every function that should run in a separate thread.
The Worker class takes care of passing the appropriate arguments to the function, so the function doesn't have to know about the big picture.
There is a caveat though: when a function needs several elements in a row it raises the ad-hoc exception KeepThemComing.